Fix zero hash (#62)
This PR fixes #60.

By _convention_ zero sized trees often define their root hash to be
`SHA256("")`, but this is not reconstructible by any means, nor used in
any consistency proof (all non zero trees are by definition consistent
with a zero sized tree). So don't try to verify the correctness of this
hash.