Remove generation iteration after writing checkpoint. (#69)
Generation numbers are only monotonically increasing and not necessarily
`[0, 1, 2, ...]` (FYI: metageneration numbers are): example
[here](https://cloud.google.com/storage/docs/object-versioning#example).

Remove code assuming the generation number of the new version after
writing a checkpoint. Practically speaking, if there is no race
condition, a client's knowledge about the checkpoint generation will
match the truth in GCS, since the client always calls ReadCheckpoint
before WriteCheckpoint. If there is a race condition, then a client's
knowledge about the checkpoint generation will be out of date, and the
write will fail because the precondition of the generation number will
not match.