Changes to Origin
The CheckpointHeaderV0 should never be used and is an artifacts of a strange couple of days from history that must not be spoken about. Origin should be unique and identify the log, thus such a constant is dangerous.

Fixed the const in the test data to use the origin string that is actually used.