Randomly create duplicate leaves
For now this creates duplicates about 10% of the time. The dupes are
always sequential in this MVP, but when we want to use this to test
deduplication in logs, we will want to add some leaves that duplicate
much older entries rather than ones that are likely to be in the same
batch to be committed.