Refactor reader/writer worker pools (#154)
These were previously stored in raw slices. Now they are stored within
the workerPool abstraction. This couples the ability to create a new
worker with the pool, making growing and shrinking the number of workers
easy.

Support w/W keys for increasing/decreasing workers

If this needs to be more granular then we can split this to different
keys for changing the size of each of the 3 pools separately.