Update token used for scorecard (#180)
This was previously using a github secret that was per-repo. I've
switched it over to an org-level secret, which should allow us to avoid
copying and pasting the secret around a number of times. Added a note
that we should switch over to branch protection rulesets instead, which
should mean that this read permission isn't needed at all.

Motivation is that the original token is expiring within the week.