Add c'tor to create ProofBuilder using only tree size (#277)
This PR adds a second c'tor for `ProofBuilder` which enables
construction of a new instance based solely on the tree size.

This provides a means to construct consistency proofs without having to
hold both checkpoints (e.g. if you're trying to convince _someone else_
of consistency between a checkpoint _they_ hold, and a newer checkpoint
_you_ hold - this is exactly what happens during the `tlog-witness`
protocol).

This is safe as generated proofs must be verified against checkpoint
root hashes by the consumer anyway; any inconsistency will be detected
at that point.